modified resources to allow method level authentication

+++ reminderresource/src/main/java/com/mojun/reminder/reminderresource/ReminderWebResource.java
@@ -7,8 +7,15 @@ import javax.ws.rs.PUT;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
+import javax.ws.rs.core.Context;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
+import javax.ws.rs.core.SecurityContext;
+
+import org.aspectj.lang.annotation.Aspect;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.stereotype.Component;
+import org.springframework.stereotype.Service;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -17,6 +24,7 @@ import com.mojun.reminder.reminderdataobj.ReminderEventList;
 import com.mojun.reminder.reminderdataobj.ReminderUser;
 import com.mojun.reminder.reminderlib.ReminderProcessor;
 
+@Service
 @Path("reminder")
 public class ReminderWebResource {
 	private static ReminderProcessor PROCESSOR = ReminderProcessor.getDefaultProcessor();
@@ -40,9 +48,12 @@ public class ReminderWebResource {
 	@Consumes({MediaType.APPLICATION_JSON})
 	@Produces({MediaType.APPLICATION_JSON})
 	@Path("/event/{userId}")
-	public Response createReminderEvent(@PathParam("userId") String userId, ReminderEvent event) {
+	public Response createReminderEvent(@PathParam("userId") String userId, ReminderEvent event, @Context SecurityContext context) {
 		if(userId == null || event == null) {
 			return Response.status(Response.Status.BAD_REQUEST).entity("UserId can not be null").build();
+		}
+		if(!context.getUserPrincipal().getName().equals(userId)) {
+			return Response.status(Response.Status.FORBIDDEN).entity(Response.Status.FORBIDDEN + "User id not match").build();
 		}
 		ReminderEvent reminderEvent = PROCESSOR.createReminderEvent(event);
 		if(reminderEvent == null) {
@@ -79,10 +90,13 @@ public class ReminderWebResource {
 	@GET
 	@Produces({MediaType.APPLICATION_JSON})
 	@Path("/event/{userId}/{reminderEventId}")
-	public Response getSingleEvent(@PathParam("userId") String userId, @PathParam("reminderEventId") String reminderEventId) {
+	public Response getSingleEvent(@PathParam("userId") String userId, @PathParam("reminderEventId") String reminderEventId, @Context SecurityContext context) {
 		
 		if(userId == null || reminderEventId == null) {
 			return Response.status(Response.Status.BAD_REQUEST).entity("userId cannot be null").build();
+		}
+		if(!context.getUserPrincipal().getName().equals(userId)) {
+			return Response.status(Response.Status.FORBIDDEN).entity(Response.Status.FORBIDDEN + "User id not match").build();
 		}
 		ReminderEvent event = PROCESSOR.getSingleReminderEvent(userId, reminderEventId);
 		if(event == null) {
@@ -99,10 +113,13 @@ public class ReminderWebResource {
 	@GET
 	@Produces({MediaType.APPLICATION_JSON})
 	@Path("/event/{userId}")
-	public Response getEventList(@PathParam("userId") String userId) {
+	public Response getEventList(@PathParam("userId") String userId, @Context SecurityContext context) {
 		
 		if(userId == null) {
 			return Response.status(Response.Status.BAD_REQUEST).entity("userId cannot be null").build();
+		}
+		if(!context.getUserPrincipal().getName().equals(userId)) {
+			return Response.status(Response.Status.FORBIDDEN).entity(Response.Status.FORBIDDEN + "User id not match").build();
 		}
 		ReminderEventList eventList = PROCESSOR.getReminderEventList(userId);
 		if(eventList == null) {
@@ -119,9 +136,12 @@ public class ReminderWebResource {
 	@GET
 	@Produces({MediaType.APPLICATION_JSON})
 	@Path("/user/{userId}")
-	public Response getUser(@PathParam("userId") String userId) {
+	public Response getUser(@PathParam("userId") String userId, @Context SecurityContext context) {
 		if(userId == null) {
 			return Response.status(Response.Status.BAD_REQUEST).entity("userId cannot be null").build();
+		}
+		if(!context.getUserPrincipal().getName().equals(userId)) {
+			return Response.status(Response.Status.FORBIDDEN).entity(Response.Status.FORBIDDEN + "User id not match").build();
 		}
 		ReminderUser user = PROCESSOR.getReminderUser(userId);
 		if(user == null) {
@@ -138,10 +158,13 @@ public class ReminderWebResource {
 	@PUT
 	@Produces({MediaType.APPLICATION_JSON})
 	@Path("/event/{userId}/{reminderEventId}")
-	public Response updateSingleEvent(@PathParam("userId") String userId, @PathParam("reminderEventId") String reminderEventId, ReminderEvent event) {
+	public Response updateSingleEvent(@PathParam("userId") String userId, @PathParam("reminderEventId") String reminderEventId, ReminderEvent event, @Context SecurityContext context) {
 		
 		if(userId == null || reminderEventId == null || event == null) {
 			return Response.status(Response.Status.BAD_REQUEST).entity("userId/reminderEventId/event cannot be null").build();
+		}
+		if(!context.getUserPrincipal().getName().equals(userId)) {
+			return Response.status(Response.Status.FORBIDDEN).entity(Response.Status.FORBIDDEN + "User id not match").build();
 		}
 		ReminderEvent reminderEvent = PROCESSOR.updateSingleEvent(event);
 		if(reminderEvent == null) {
@@ -158,10 +181,13 @@ public class ReminderWebResource {
 	@Produces({MediaType.APPLICATION_JSON})
 	@Consumes({MediaType.APPLICATION_JSON})
 	@Path("/user/{userId}") 
-	public Response updateReminderUser(@PathParam("userId") String userId, ReminderUser user) {
+	public Response updateReminderUser(@PathParam("userId") String userId, ReminderUser user, @Context SecurityContext context) {
 		
 		if(userId == null || user == null) {
 			return Response.status(Response.Status.BAD_REQUEST).entity("userId/user cannot be null").build();
+		}
+		if(!context.getUserPrincipal().getName().equals(userId)) {
+			return Response.status(Response.Status.FORBIDDEN).entity(Response.Status.FORBIDDEN + "User id not match").build();
 		}
 		ReminderUser reminderUser = PROCESSOR.updateReminderUser(user);
 		if(reminderUser == null) {
@@ -177,9 +203,12 @@ public class ReminderWebResource {
 	@DELETE
 	@Produces({MediaType.APPLICATION_JSON})
 	@Path("/event/{userId}/{reminderEventId}")
-	public Response deleteSingleEvent (@PathParam("userId") String userId, @PathParam("reminderEventId") String reminderEventId) {
+	public Response deleteSingleEvent (@PathParam("userId") String userId, @PathParam("reminderEventId") String reminderEventId, @Context SecurityContext context) {
 		if(userId == null || reminderEventId == null) {
 			return Response.status(Response.Status.BAD_REQUEST).entity("userId/reminderEventId cannot be null").build();
+		}
+		if(!context.getUserPrincipal().getName().equals(userId)) {
+			return Response.status(Response.Status.FORBIDDEN).entity(Response.Status.FORBIDDEN + "User id not match").build();
 		}
 		ReminderEvent reminderEvent = PROCESSOR.deleteSingleEvent(userId, reminderEventId);
 		if(reminderEvent != null) {
@@ -191,9 +220,12 @@ public class ReminderWebResource {
 	@DELETE
 	@Produces({MediaType.APPLICATION_JSON})
 	@Path("/user/{userId}")
-	public Response deleteUser(@PathParam("userId") String userId) {
+	public Response deleteUser(@PathParam("userId") String userId, @Context SecurityContext context) {
 		if(userId == null) {
 			return Response.status(Response.Status.BAD_REQUEST).entity("userId cannot be null").build();
+		}
+		if(!context.getUserPrincipal().getName().equals(userId)) {
+			return Response.status(Response.Status.FORBIDDEN).entity(Response.Status.FORBIDDEN + "User id not match").build();
 		}
 		ReminderUser reminderUser = PROCESSOR.deleteUser(userId);
 		if(reminderUser != null) {
